agentic_version/requirements.txt
# Core agent framework - using compatible versions
langgraph>=0.1.0
langchain>=0.1.0
langchain-openai>=0.0.5
langchain-community>=0.0.10

# Database and data processing
psycopg2-binary>=2.9.0
sqlalchemy>=2.0.0
pandas>=2.0.0
numpy>=1.24.0

# AI and embeddings
openai>=1.3.0
tiktoken>=0.5.0
faiss-cpu>=1.7.0
scikit-learn>=1.3.0

# NLP and text processing
nltk>=3.8.0
spacy>=3.7.0

# Configuration and utilities
python-dotenv>=1.0.0
pydantic>=2.0.0

# Data visualization
plotly>=5.17.0
matplotlib>=3.8.0
seaborn>=0.13.0

# Web framework (optional, for API endpoints)
fastapi>=0.104.0
uvicorn>=0.24.0

# Additional utilities
requests>=2.31.0
python-multipart>=0.0.6
